HOME

HOME

QTool.Hamiltonian — Type.

Object to hold general time dependent Hamiltonians, whose form is assumed to be a summation of time dependent function times constant matrices.

source

QTool.Hamiltonian — Method.

Evaluate the Hamiltonian at time t

source

QTool.comm! — Method.

comm!(Y, A, B)

Calculate the commutator of matrices A and B. Write the result in Y

source

QTool.comm — Method.

comm(A, B)

Calculate the commutator of matrices A and B

source

QTool.construct_hamming_weight_op — Method.

construct_hamming_weight_op(num_qubit::Int64, op::String)

Construct the Hamming weight operator for system of size num_qubit. The type of the Hamming weight operator is specified by op: "x", "y" or "z".

Examples

julia> construct_hamming_weight_op(2,"z")
4×4 Array{Complex{Float64},2}:
 0.0+0.0im  0.0+0.0im  0.0+0.0im  0.0+0.0im
 0.0+0.0im  1.0+0.0im  0.0+0.0im  0.0+0.0im
 0.0+0.0im  0.0+0.0im  1.0+0.0im  0.0+0.0im
 0.0+0.0im  0.0+0.0im  0.0+0.0im  2.0+0.0im

source

QTool.eigen_state_eval — Method.

eigen_state_eval(hfun, t[, levels=[1,]])

Calculate the eigen states of Hamiltonian hfun at each points of t. The output levels are specified by levels argument.

source

QTool.eigen_value_eval — Method.

eigen_value_eval(hfun, t[, levels])

Calculate the eigen values of Hamiltonian hfun at each points of t. The output levels are specified by levels argument.

source

QTool.inst_population — Method.

function inst_population(t, states, hamiltonian; level=1)

For a time series quantum states given by states, whose time points are given by t, calculate the population of instantaneous eigenstates of hamiltonian. The levels of the instantaneous eigenstates are specified by level, which can be any slice index.

source

QTool.matrix_decompose — Method.

matrix_decompose(mat::Array{T,2}, basis::Array{Array{T,2},1})

Decompse matrix mat onto matrix basis basis

Examples

julia> matrix_decompose(1.0*σx+2.0*σy+3.0*σz, [σx,σy,σz])
3-element Array{Complex{Float64},1}:
 1.0 + 0.0im
 2.0 + 0.0im
 3.0 + 0.0im

source

QTool.q_translate — Method.

q_translate(h::String)

Convert a string h representing multi-qubits Pauli matrices summation into its numerical form.

Examples

julia> q_translate("X+2.0Z")
2×2 Array{Complex{Float64},2}:
 2.0+0.0im   1.0+0.0im
 1.0+0.0im  -2.0+0.0im

source

QTool.temperature_2_beta — Method.

temperature_2_beta(T; unit=:h)

Convert temperature from mK to unit*β

source

QTool.temperature_2_freq — Method.

temperature_2_freq(T)

Convert temperature from mK to GHz(physical unit)

source