Projection

Projection Tools

QuantumAnnealingTools.LowLevelSystem — Type.

LowLevelSystem

Object for a projected low level system. The projection is only valid for real Hamiltonians.

Fields

s – unitless time grid.

ev – energy values for different levels.

dθ – geometric terms.

op – projected system bath interaction operators

ref – energy eigenstates at the final time

lvl – number of levels being kept.

source

QuantumAnnealingTools.get_dθ — Function.

get_dθ(sys::LowLevelSystem, i=1, j=2)

Get the geometric terms between i, j energy levels from LowLevelSystem.

source

QuantumAnnealingTools.RotatedTwoLevelSystem — Type.

RotatedTwoLevelSystem

Object for a rotated two level system.

Fields

s – unitless time grid.

ω – ω₁₂.

T – off diagonal element of the Hamiltonian.

G – geometric term.

a, b, c, d – projected system bath parameters

θ – rotation angle

source

QuantumAnnealingTools.proj_low_lvl — Method.

proj_low_lvl(hfun, dhfun, interaction, s_axis::AbstractArray{T, 1}; ref=nothing, tol=1e-4, lvl=2) where T<:Number

Project a multi-level quantum system to the lowest lvl levels. hfun and dhfun are the Hamiltonian and its derivative. interaction is a list of system bath interaction operators. s_axis is the grid for the computation. ref is the reference for the initial eigenstates. tol specify the error tolerance for eigen-decomposition if sparse matrix is used.

source