QTBase.PauliVec — Constant

PauliVec

Constants for the eigenvectors of single qubit Pauli matrices. Indices 1, 2 and 3 corresponds to the eigenvectors of $σ_x$, $σ_y$ and $σ_z$.

Examples

julia> σx*PauliVec[1][1] == PauliVec[1][1]
true

QTBase.AbstractAnnealing — Type

abstract type AbstractAnnealing

Base for types defining quantum annealing process.

QTBase.AbstractBath — Type

abstract type AbstractBath

Base for types defining bath object.

QTBase.AbstractCouplings — Type

abstract type AbstractCouplings

Base for types defining system bath coupling operators in open quantum system models.

QTBase.AbstractDenseHamiltonian — Type

abstract type AbstractDenseHamiltonian <: AbstractHamiltonian{T<:Number}

Base for types defining Hamiltonians using dense matrices.

QTBase.AbstractHamiltonian — Type

abstract type AbstractHamiltonian

Suptertype for Hamiltonians with elements of type T. Any Hamiltonian object should implement two interfaces: H(t) and H(du, u, p, t).

QTBase.AbstractLiouvillian — Type

abstract type AbstractLiouvillian

QTBase.AbstractSparseHamiltonian — Type

abstract type AbstractSparseHamiltonian <: AbstractHamiltonian{T<:Number}

Base for types defining Hamiltonians using sparse matrices.

QTBase.AdiabaticFrameHamiltonian — Type

struct AdiabaticFrameHamiltonian{T} <: AbstractDenseHamiltonian{T}

Defines a time dependent Hamiltonian in adiabatic frame.

Fields

geometric

Geometric part

diagonal

Adiabatic part

size

Size of the Hamiltonian

QTBase.AdiabaticFrameHamiltonian — Method

function AdiabaticFrameHamiltonian(ωfuns, geofuns)

Constructor of adiabatic frame Hamiltonian. ωfuns is a 1-D array of functions which specify the eigen energies (in GHz) of the Hamiltonian. geofuns is a 1-D array of functions which specifies the geometric phases of the Hamiltonian. geofuns can be thought as a flattened lower triangular matrix (without diagonal elements) in column-major order.

QTBase.Annealing — Type

mutable struct Annealing{hType, uType} <: AbstractAnnealing{hType,uType}

Defines a quantum annealing process.

Fields

H

Hamiltonian for the annealing.

u0

Initial state for the annealing.

annealing_parameter

Function of annealing parameter s wrt to t

interactions

A system bath interaction set.

QTBase.ConstantCouplings — Type

struct ConstantCouplings <: AbstractCouplings

Defines constant system bath coupling operators.

Fields

mats

1-D array for independent coupling operators

str_rep

String representation for the coupling (for display purpose)

QTBase.ConstantCouplings — Method

function ConstantCouplings(c::Vector{T}; sp = false, unit=:h) where T <: AbstractString

If the first argument is a 1-D array of strings. The constructor will automatically construct the matrics represented by the string representations.

QTBase.ConstantCouplings — Method

function ConstantCouplings(mats; str_rep=nothing, unit=:h)

Constructor of ConstantCouplings object. mats is 1-D array of matrices. str_rep is the optional string representation of the coupling terms. unit is the unit one – :h or :ħ. The mats will be scaled by $2π$ is unit is :h.

QTBase.CustomBath — Type

mutable struct CustomBath <: AbstractBath

An custum bath object defined by the two-point correlation function and the corresponding spectrum.

cfun

correlation function

γ

spectrum

QTBase.CustomCouplings — Type

struct CustomCouplings <: AbstractTimeDependentCouplings

CustomCouplings is a container for any user defined coupling operators.

Fields

coupling

A 1-D array of callable objects that returns coupling matrices

size

Size of the coupling operator

QTBase.CustomDenseHamiltonian — Type

struct CustomDenseHamiltonian{T<:Number, in_place} <: AbstractDenseHamiltonian{T<:Number}

Defines a time dependent Hamiltonian object with custom function.

Fields

f

Function for the Hamiltonian H(s)

u_cache

Internal cache

size

Size

QTBase.DaviesGenerator — Type

struct DaviesGenerator <: AbstractLiouvillian

Defines Davies generator

Fields

coupling

System bath coupling operators

γ

Spectrum density

S

Lambshift spectrum density

QTBase.DenseHamiltonian — Type

struct DenseHamiltonian{T<:Number} <: AbstractDenseHamiltonian{T<:Number}

Defines a time dependent Hamiltonian object with dense Matrices.

Fields

f

List of time dependent functions

m

List of constant matrices

u_cache

Internal cache

size

Size

EIGS

Eigen decomposition routine

QTBase.DenseHamiltonian — Method

DenseHamiltonian(funcs, mats; unit, EIGS)

Constructor of DenseHamiltonian object. funcs and mats are a list of time dependent functions and the corresponding matrices. unit is the unit one – :h or :ħ. The mats will be scaled by $2π$ if unit is :h.

EIGS is the initializer for the eigen decomposition routine for the Hamiltonian. It should return a function of signature: (H, s, lvl) -> (w, v). The default method EIGEN_DEFAULT will use LAPACK routine.

QTBase.DenseHamiltonian — Method

function (h::DenseHamiltonian)(s::Real)

Calling the Hamiltonian returns the value $2πH(s)$. The argument s is in the unitless time. The returned matrix is in angular frequency.

QTBase.EnsembleFluctuator — Type

struct EnsembleFluctuator{T} <: AbstractBath

An ensemble of random telegraph noise.

f

A list of RTNs

QTBase.Fluctuators — Type

mutable struct Fluctuators <: AbstractLiouvillian

Defines a fluctuator ensemble controller

Fields

coupling

system-bath coupling operator

dist

waitting time distribution for every fluctuators

b0

cache for each fluctuator value

next_idx

index of the fluctuator to be flipped next

next_τ

time interval for next flip event

n

noise value

QTBase.InplaceUnitary — Type

struct InplaceUnitary

A tag for inplace unitary function. func(cache, t) is the actual inplace update function.

Fields

func

inplace update function

QTBase.Interaction — Type

struct Interaction

An object to hold system operator and the corresponding bath object.

coupling

system operator

bath

bath coupling to the system operator

QTBase.InteractionSet — Type

struct InteractionSet{T<:Tuple}

An container for different system-bath interactions.

interactions

A tuple of Interaction

QTBase.InterpDenseHamiltonian — Type

struct InterpDenseHamiltonian{T} <: AbstractDenseHamiltonian{T}

Defines interpolating DenseHamiltonian object

Fields

interp_obj

Interpolating object

size

Size

QTBase.InterpSparseHamiltonian — Type

struct InterpSparseHamiltonian{T} <: AbstractSparseHamiltonian{T}

Defines interpolating SparseHamiltonian object

Fields

interp_obj

Interpolating object

size

Size

QTBase.ODEParams — Type

struct ODEParams

Defines a complete set of ODE parameters, which includes Hamiltonian, total annealing time, open system and control objects.

Fields

L

Hamiltonian

tf

Total annealing time

annealing_parameter

Function to convert physical time to annealing parameter

control

Annealing control object

QTBase.OhmicBath — Type

OhmicBath

Ohmic bath object to hold a particular parameter set.

Fields

η – strength.

ωc – cutoff frequence.

β – inverse temperature.

QTBase.OpenSysOp — Type

struct OpenSysOp{diagonalization, adiabatic_frame}

Defines an open system DiffEqOperator in the specific basis.

Fields

H

Hamiltonian

opensys

Open system part

lvl

Levels to truncate

u_cache

Internal cache

QTBase.ProjectedCoupling — Type

struct ProjectedCoupling

Object for a projected coupling parameterized by NIBA parameters. This object holds the numerical values at gridded points.

Fields

s

Time grid (unitless) for projection

a

$(σ_m - σ_n)^2$

b

$|σ_{mn}|^2$

c

$σ_{mn}(σ_m-σ_n)$

d

$σ_{mn}(σ_m+σ_n)$

QTBase.ProjectedSystem — Type

struct ProjectedSystem

Object for a projected low level system. The projection is only valid for real Hamiltonians.

Fields

s

Time grid (unitless) for projection

ev

Energy values for different levels

dθ

Geometric terms

op

Projected system bath interaction operators

ref

Energy eigenstates at the final time

lvl

Number of levels being kept

QTBase.ProjectedTG — Type

struct ProjectedTG

Object for a projected system Hamiltonian parameterized by NIBA parameters. This object holds the numerical values at gridded points.

Fields

s

Time grid (unitless) for projection

ω

Frequencies of given basis states

T

Adiabatic part

G

Geometric phase

QTBase.RedfieldGenerator — Type

struct RedfieldGenerator <: AbstractLiouvillian

Defines RedfieldGenerator.

Fields

ops

system-bath coupling operator

unitary

close system unitary

cfun

bath correlation function

atol

absolute error tolerance for integration

rtol

relative error tolerance for integration

Ut

cache matrix for inplace unitary

Uτ

cache matrix for inplace unitary

Λ

cache matrix for integration

Ta

tf minus coarse grain time scale

QTBase.SparseHamiltonian — Type

struct SparseHamiltonian{T<:Number} <: AbstractSparseHamiltonian{T<:Number}

Defines a time dependent Hamiltonian object with sparse Matrices.

Fields

f

List of time dependent functions

m

List of constant matrices

u_cache

Internal cache

size

Size

EIGS

Eigen decomposition routine

QTBase.SparseHamiltonian — Method

SparseHamiltonian(funcs, mats; unit, EIGS)

Constructor of SparseHamiltonian object. funcs and mats are a list of time dependent functions and the corresponding matrices. unit is the unit one – :h or :ħ. The mats will be scaled by $2π$ is unit is :h.

EIGS is the initializer for the eigen decomposition routine for the Hamiltonian. It should return a function of signature: (H, s, lvl) -> (w, v). The default method EIGEN_DEFAULT will use LAPACK routine.

QTBase.SparseHamiltonian — Method

function (h::SparseHamiltonian)(t::Real)

Calling the Hamiltonian returns the value $2πH(t)$.

QTBase.TimeDependentCoupling — Type

struct TimeDependentCoupling

Defines a single time dependent system bath coupling operator. It is defined as $S(s)=∑f(s)×M$. Keyword argument unit set the unit one – $h$ or $ħ$.

Fields

funcs

1-D array of time dependent functions

mats

1-D array of constant matrics

QTBase.TimeDependentCouplings — Type

struct TimeDependentCouplings <: AbstractTimeDependentCouplings

Defines an 1-D array of time dependent system bath coupling operators.

Fields

coupling

A tuple of single TimeDependentCoupling operators

QTBase.ΓMatrix — Type

struct ΓMatrix

Defines rates of Davies form ME acting on the diagonal elements.

Fields

m

interpolated Γ matrix

lvl

levels of the system

QTBase.:⊗ — Function

⊗(A, B)

Calculate the tensor product of A and B.

Examples

julia> σx⊗σz
4×4 Array{Complex{Float64},2}:
 0.0+0.0im   0.0+0.0im  1.0+0.0im   0.0+0.0im
 0.0+0.0im  -0.0+0.0im  0.0+0.0im  -1.0+0.0im
 1.0+0.0im   0.0+0.0im  0.0+0.0im   0.0+0.0im
 0.0+0.0im  -1.0+0.0im  0.0+0.0im  -0.0+0.0im

QTBase.EIGEN_DEFAULT — Method

function EIGEN_DEFAULT(H)

The default initializer for eigen factorization method. It returns a function of signature: (H, s, lvl) -> (w, v). H is the Hamiltonian object, s is the dimensionless time and lvl is the energy levels to keep. This default initializer will use LAPACK routine for both dense and sparse matrices.

QTBase.EIGEN_DEFAULT — Method

function EIGEN_DEFAULT(u_cache)

The default initializer for eigen factorization method. It returns a function of signature: (H, s, lvl) -> (w, v). u_cache is the cache for Hamiltonian object, s is the time argument for the Hamiltonian and lvl is the energy levels to keep. This default initializer will use LAPACK routine for both dense and sparse matrices.

QTBase.Ohmic — Method

Ohmic(η, fc, T)

Construct OhmicBath from parameters with physical unit: η–unitless interaction strength; fc–cutoff frequency in GHz; T–temperature in mK.

QTBase.S — Method

S(w, bath; atol)

Calculate the Lamb shift of Ohmic spectrum. atol is the absolute tolerance for Cauchy principal value integral.

QTBase.ame_jump — Method

ame_jump(A::OpenSysOp, u, p, t::Real)

Calculate the jump operator for the OpenSysOp at time t.

QTBase.check_positivity — Method

check_positivity(m)

Check if matrix m is positive. Return true is the minimum eigenvalue of m is greater than or equal to 0.

QTBase.check_unitary — Method

check_unitary(𝐔; rtol, atol)

Test if 𝐔 is a unitary matrix. The function checks how close both $𝐔𝐔^†$ and $𝐔^†𝐔$ are to $I$, with relative and absolute error given by rtol, atol.

Examples

julia> check_unitary(exp(-1.0im*5*0.5*σx))
true

QTBase.coarse_grain_timescale — Method

coarse_grain_timescale(bath, lim; rtol, atol)

Calculate the optimal coarse grain time scale $T_a$ for a total evolution time lim. atol and rtol are the absolute and relative error for the integration.

QTBase.collective_coupling — Method

function collective_coupling(op, num_qubit; sp=false)

Create ConstantCouplings object with operator op on each qubits. op is the string representation of one of the Pauli matrices. num_qubit is the total number of qubits. sp set whether to use sparse matrices. unit set the unit one – $h$ or $ħ$.

QTBase.collective_operator — Method

collective_operator(op, num_qubit; sp=false)

Construct the collective operator for a system of num_qubit qubits. op is the name of the collective Pauli matrix. For example, the following code construct an $IZ + ZI$ matrix. Generate sparse matrix when sp is set to true.

Examples

julia> collective_operator("z", 2)
4×4 Array{Complex{Float64},2}:
 2.0+0.0im  0.0+0.0im  0.0+0.0im   0.0+0.0im
 0.0+0.0im  0.0+0.0im  0.0+0.0im   0.0+0.0im
 0.0+0.0im  0.0+0.0im  0.0+0.0im   0.0+0.0im
 0.0+0.0im  0.0+0.0im  0.0+0.0im  -2.0+0.0im

QTBase.concatenate — Method

function concatenate(args...)

Concatenate multiple ProjectedSystem objects into a single one. The arguments need to be in time order. The ref field of the new object will have the same value as the last input arguments.

QTBase.construct_interpolations — Method

function construct_interpolations(x, y; method = "BSpline", order = 2, extrapolation = "line")

Construct interpolation of N-D array y along its last dimension on grid x. method specifies the interpolation algorithm, which can be either "BSpline" or "Gridded". If x is an AbstractRange, the default method is "BSpline", otherwise the default method is "Gridded". order is the interpolation order. If x is AbstractRange, the default order is 2, otherwise the default order is 1. extrapolation specifies the extrapolation methods.

QTBase.eigen_decomp — Method

eigen_decomp(H, s; lvl)

Calculate the eigen value decomposition of the Hamiltonian H at an array of time points s. The output keeps the lowest lvl eigenstates and their corresponding eigenvalues. Output (vals, vecs) have the dimensions of (lvl, length(s)) and (size(H, 1), lvl, length(s)) respectively.

QTBase.eigen_decomp — Method

eigen_decomp(H, s; lvl)

Calculate the eigen value decomposition of the Hamiltonian H at time s. Keyword argument lvl specifies the number of levels to keep in the output. w is a vector of eigenvalues and v is a matrix of the eigenvectors in the columns. (The kth eigenvector can be obtained from the slice v[:, k].) w will be in unit of GHz.

eig_init is the initializer for eigen factorization routine. It returns a function of signature: (H, s, lvl) -> (w, v). The default initializer EIGEN_DEFAULT will use LAPACK routine for both dense and sparse matrices.

QTBase.evaluate — Method

evaluate(H::AbstractHamiltonian, s)

Evaluate the time dependent Hamiltonian at time s with the unit of GHz

QTBase.evaluate — Method

function evaluate(H::AdiabaticFrameHamiltonian, s, tf)

Evaluate the adiabatic frame Hamiltonian at (unitless) time s, with total annealing time tf (in the unit of $ns$). The final result is given in unit of $GHz$.

QTBase.freq_2_temperature — Method

freq_2_temperature(freq)

Convert frequency in GHz to temperature in mK.

QTBase.get_dθ — Function

get_dθ(sys::ProjectedSystem, i=1, j=2)

Get the geometric terms between i, j energy levels from ProjectedSystem.

QTBase.gibbs_state — Method

gibbs_state(h, β)

Calculate the Gibbs state of the matrix h at temperature T (mK).

Examples

julia> gibbs_state(σz, 10)
2×2 Array{Complex{Float64},2}:
 0.178338+0.0im       0.0+0.0im
      0.0+0.0im  0.821662+0.0im

QTBase.gradient — Method

function gradient(itp, s)

Calculate the gradient of itp at s.

QTBase.hamming_weight_operator — Method

hamming_weight_operator(num_qubit::Int64, op::String; sp=false)

Construct the Hamming weight operator for system of size num_qubit. The type of the Hamming weight operator is specified by op: "x", "y" or "z". Generate sparse matrix when sp is set to true.

Examples

julia> hamming_weight_operator(2,"z")
4×4 Array{Complex{Float64},2}:
 0.0+0.0im  0.0+0.0im  0.0+0.0im  0.0+0.0im
 0.0+0.0im  1.0+0.0im  0.0+0.0im  0.0+0.0im
 0.0+0.0im  0.0+0.0im  1.0+0.0im  0.0+0.0im
 0.0+0.0im  0.0+0.0im  0.0+0.0im  2.0+0.0im

QTBase.inst_population — Method

inst_population(s, states, H; lvl)

For a time series quantum states given by states, whose time points are given by s, calculate the population of instantaneous eigenstates of H. The levels of the instantaneous eigenstates are specified by lvl, which can be any slice index.

eig_init is the initializer for eigen factorization routine. It returns a function of signature: (H, s, lvl) -> (w, v). The default initializer EIGEN_DEFAULT will use LAPACK routine for both dense and sparse matrices.

QTBase.local_field_term — Method

local_field_term(h, idx, num_qubit; sp=false)

Construct local Hamiltonian of the form $∑hᵢσᵢᶻ$. idx is the index of all local field terms and h is a list of the corresponding weights. num_qubit is the total number of qubits. Generate sparse matrix when sp is set to true.

Examples

julia> local_field_term([1.0, 0.5], [1, 2], 2) == σz⊗σi+0.5σi⊗σz
true

QTBase.log_uniform — Method

log_uniform(a, b, num; base)

Generate a log-uniformly distributed array with num elements between a and b. The base of log is base with default value 10.

QTBase.low_level_matrix — Method

low_level_matrix(M, lvl)

Calculate the matrix M projected to lower energy subspace containing lvl energy lvl.

Examples

julia> low_level_matrix(σx⊗σx, 2)
4×4 Array{Complex{Float64},2}:
 -0.5+0.0im   0.0+0.0im   0.0+0.0im   0.5+0.0im
  0.0+0.0im  -0.5+0.0im   0.5+0.0im   0.0+0.0im
  0.0+0.0im   0.5+0.0im  -0.5+0.0im   0.0+0.0im
  0.5+0.0im   0.0+0.0im   0.0+0.0im  -0.5+0.0im

QTBase.matrix_decompose — Method

matrix_decompose(mat, basis)

Decompse matrix mat onto matrix basis basis

Examples

julia> matrix_decompose(1.0*σx+2.0*σy+3.0*σz, [σx,σy,σz])
3-element Array{Complex{Float64},1}:
 1.0 + 0.0im
 2.0 + 0.0im
 3.0 + 0.0im

QTBase.partial_trace — Method

partial_trace(ρ, qubit_2_keep)

Calculate the partial trace of the density matrix ρ. qubit_2_keep is an array of qubit indices to keep.

Examples

julia> ρ1 = [0.4 0.5; 0.5 0.6]; ρ2 = [0.5 0; 0 0.5];
julia> partial_trace(ρ1⊗ρ2, [1])
2×2 Array{Float64,2}:
 0.4  0.5
 0.5  0.6

QTBase.project_to_lowlevel — Method

project_to_lowlevel(H::AbstractHamiltonian{T}, dH, coupling, s_axis; lvl=2, eig_init = EIGEN_DEFAULT)

Project a Hamiltonian H to the lowest lvl level subspace. dH is the derivative of Hamiltonian and coupling is the system-bath interaction operator. They should be callable with a single argument – annealing parameter s. s_axis is the (unitless) times to calculate the projection.

QTBase.q_translate — Method

q_translate(h::String)

Convert a string h representing multi-qubits Pauli matrices summation into its numerical form.

Examples

julia> q_translate("X+2.0Z")
2×2 Array{Complex{Float64},2}:
 2.0+0.0im   1.0+0.0im
 1.0+0.0im  -2.0+0.0im

QTBase.q_translate_state — Method

q_translate_state(h::String; normal=false)

Convert a string representation of quantum state to a vector. The keyword argument normal indicates whether to normalize the output vector. (Currently only '0' and '1' are supported)

Examples

Single term:

julia> q_translate_state("001")
8-element Array{Complex{Float64},1}:
 0.0 + 0.0im
 1.0 + 0.0im
 0.0 + 0.0im
 0.0 + 0.0im
 0.0 + 0.0im
 0.0 + 0.0im
 0.0 + 0.0im
 0.0 + 0.0im

Multiple terms:

julia> q_translate_state("(101)+(001)", normal=true)
8-element Array{Complex{Float64},1}:
                0.0 + 0.0im
 0.7071067811865475 + 0.0im
                0.0 + 0.0im
                0.0 + 0.0im
                0.0 + 0.0im
 0.7071067811865475 + 0.0im
                0.0 + 0.0im
                0.0 + 0.0im

QTBase.single_clause — Method

single_clause(ops, q_ind, weight, num_qubit; sp=false)

Construct a single clause of the multi-qubits Hamiltonian. ops is a list of Pauli operator names which appears in this clause. q_ind is the list of indices corresponding to the Pauli matrices in ops. weight is the constant factor of this clause. num_qubit is the total number of qubits. A sparse matrix can be construct by setting sp to true. The following example construct a clause of $Z_1 I Z_3/2$.

Examples

julia> single_clause(["z", "z"], [1, 3], 0.5, 3)
8×8 Array{Complex{Float64},2}:
 0.5+0.0im   0.0+0.0im  0.0+0.0im   0.0+0.0im   0.0+0.0im   0.0+0.0im   0.0+0.0im   0.0+0.0im
 0.0+0.0im  -0.5+0.0im  0.0+0.0im  -0.0+0.0im   0.0+0.0im  -0.0+0.0im   0.0+0.0im  -0.0+0.0im
 0.0+0.0im   0.0+0.0im  0.5+0.0im   0.0+0.0im   0.0+0.0im   0.0+0.0im   0.0+0.0im   0.0+0.0im
 0.0+0.0im  -0.0+0.0im  0.0+0.0im  -0.5+0.0im   0.0+0.0im  -0.0+0.0im   0.0+0.0im  -0.0+0.0im
 0.0+0.0im   0.0+0.0im  0.0+0.0im   0.0+0.0im  -0.5+0.0im  -0.0+0.0im  -0.0+0.0im  -0.0+0.0im
 0.0+0.0im  -0.0+0.0im  0.0+0.0im  -0.0+0.0im  -0.0+0.0im   0.5-0.0im  -0.0+0.0im   0.0-0.0im
 0.0+0.0im   0.0+0.0im  0.0+0.0im   0.0+0.0im  -0.0+0.0im  -0.0+0.0im  -0.5+0.0im  -0.0+0.0im
 0.0+0.0im  -0.0+0.0im  0.0+0.0im  -0.0+0.0im  -0.0+0.0im   0.0-0.0im  -0.0+0.0im   0.5-0.0im

QTBase.standard_driver — Method

standard_driver(num_qubit; sp=false)

Construct the standard driver Hamiltonian for a system of num_qubit qubits. For example, a two qubits standard driver matrix is $IX + XI$. Generate sparse matrix when sp is set to true.

QTBase.temperature_2_freq — Method

temperature_2_freq(T)

Convert temperature from mK to GHz.

QTBase.temperature_2_β — Method

 temperature_2_β(T)

Convert physical temperature T in mK to thermodynamic β in the unit of inverse angular frequency, that is to say $β = ħ/kT$.

QTBase.two_local_term — Method

two_local_term(j, idx, num_qubit; sp=false)

Construct local Hamiltonian of the form $∑Jᵢⱼσᵢᶻσⱼᶻ$. idx is the index of all two local terms and j is a list of the corresponding weights. num_qubit is the total number of qubits. Generate sparse matrix when sp is set to true.

Examples

julia> two_local_term([1.0, 0.5], [[1,2], [1,3]], 3) == σz⊗σz⊗σi + 0.5σz⊗σi⊗σz
true

QTBase.β_2_temperature — Method

β_2_temperature(β)

Convert thermodynamic β in the unit of inverse angular frequency to physical temperature T in GHz.

QTBase.γ — Method

γ(ω, params)

Calculate Ohmic spectrum density, defined as a full Fourier transform on the bath correlation function.

QTBase.τ_B — Method

τ_B(cfun, lim, τsb; rtol, atol)

Calculate the bath correlation time $τ_B$. The upper limit lim and τsb need to be manually specified. atol and rtol are the absolute and relative error for the integration.

QTBase.τ_SB — Method

τ_SB(cfun; lim, rtol, atol)

Calculate $τ_{SB}$ from the bath correlation function cfun. It is defined as the integration of the absolute value of bath correlation function from zero to lim. The default value of lim is Inf. atol and rtol are the absolute and relative error of the integration.