Callbacks

OpenQuantumTools.InstPulseCallback — Method

InstPulseCallback(tstops, pulse_update)

Builds the callback to apply instantaneous pulses during the evolution.

...

Arguments

tstops: time points to apply pulses.

pulse_update: udpate function for the state of the system. It taks two argument pulse_update(c, i) where c is the state vector or density matrix and i is the index of the pulse being applied.

...

OpenQuantumTools.PositivityCheckCallback — Method

PositivityCheckCallback()

Builds the callback to check positivity of the system density matrix during the evolution.