Solvers

High level interface

Closed System

QuantumAnnealingTools.solve_schrodinger — Method

solve_schrodinger(A, tf; tspan, kwargs...)

Solve Schrodinger dynamics defined by Annealing object A.

...

Arguments

A::Annealing: the Annealing object.

tf::Real: the total annealing time.

tspan = (0, tf): time interval to solve.

kwargs: other keyword arguments supported by DifferentialEquations.jl.

...

QuantumAnnealingTools.solve_unitary — Method

solve_unitary(A, tf; vectorize, tspan, kwargs...)

Solve the unitary of Annealing object A.

...

Arguments

A::Annealing: the Annealing object.

tf::Real: the total annealing time.

vectorize::Bool = false: whether to vectorize the density matrix.

tspan = (0, tf): time interval to solve.

kwargs: other keyword arguments supported by DifferentialEquations.jl.

...

QuantumAnnealingTools.solve_von_neumann — Method

solve_von_neumann(A, tf; tspan, vectorize, kwargs...)

Solve the Von Neumann equation of Annealing object A.

...

Arguments

A::Annealing: the Annealing object.

tf::Real: the total annealing time.

vectorize::Bool = false: whether to vectorize the density matrix.

tspan = (0, tf): time interval to solve.

kwargs: other keyword arguments supported by DifferentialEquations.jl.

...

Open System

QuantumAnnealingTools.solve_redfield — Method

solve_redfield(A, tf, unitary; vectorize, int_atol, int_rtol, Ta, kwargs...)

Solve the time dependent Redfield equation for Annealing defined by A with total annealing time tf.

...

Arguments

A::Annealing: the Annealing object.

tf::Real: the total annealing time.

unitary: precalculated unitary of close system evolution.

vectorize::Bool = false: whether to vectorize the density matrix.

int_atol = 1e-8: the absolute error tolerance for integration.

int_rtol = 1e-6: the relative error tolerance for integration.

Ta = tf: the time scale for backward integration.

kwargs: other keyword arguments supported by DifferentialEquations.jl.

...

QuantumAnnealingTools.solve_CGME — Method

solve_CGME(A, tf, unitary; vectorize, Ta, int_atol, int_rtol, kwargs...)

Solve the time dependent CGME for Annealing defined by A with total annealing time tf.

...

Arguments

A::Annealing: the Annealing object.

tf::Real: the total annealing time.

unitary: precalculated unitary of close system evolution.

vectorize::Bool = false: whether to vectorize the density matrix.

Ta = nothing: coarse-graining time. If set to nothing, the solver will automatically choose the value.

int_atol = 1e-8: the absolute error tolerance for integration.

int_rtol = 1e-6: the relative error tolerance for integration.

kwargs: other keyword arguments supported by DifferentialEquations.jl.

...

QuantumAnnealingTools.solve_ule — Method

solve_ule(A, tf, unitary; vectorize, int_atol, int_rtol, Ta, kwargs...)

Solve the time dependent ULE for Annealing defined by A with total annealing time tf.

...

Arguments

A::Annealing: the Annealing object.

tf::Real: the total annealing time.

unitary: precalculated unitary of close system evolution.

vectorize::Bool = false: whether to vectorize the density matrix.

int_atol = 1e-8: the absolute error tolerance for integration.

int_rtol = 1e-6: the relative error tolerance for integration.

Ta = tf: the time scale for integration region.

kwargs: other keyword arguments supported by DifferentialEquations.jl.

...

QuantumAnnealingTools.solve_ame — Method

solve_ame(A, tf; tspan, ω_hint, lambshift, lvl, vectorize, kwargs...)

Solve the adiabatic master equation for Annealing defined by A with total annealing time tf.

...

Arguments

A::Annealing: the Annealing object.

tf::Real: the total annealing time.

tspan = (0, tf): time interval to solve.

ω_hint=[] : grid for precalculating the lambshift; skip the precalculation if empty.

'lambshift::Bool=true' : whether to include Lambshift in the calculation.

lvl::Int=size(A.H, 1) : number of levels to keep. The default value is the dimension for the Hamiltonian.

vectorize::Bool = false: whether to vectorize the density matrix.

kwargs : other keyword arguments supported by DifferentialEquations.jl.

...