Bath Module

Ohmic Bath Object

QuantumAnnealingTools.OhmicBath — Type

OhmicBath

Ohmic bath object to hold a particular parameter set.

Fields

η – strength.

ωc – cutoff frequence.

β – inverse temperature.

QuantumAnnealingTools.Ohmic — Method

Ohmic(η, fc, T)

Construct OhmicBath from parameters with physical unit: η–unitless interaction strength; fc–cutoff frequency in GHz; T–temperature in mK.

QuantumAnnealingTools.correlation — Method

correlation(τ, bath)

Calculate the correlation function of bath.

QuantumAnnealingTools.polaron_correlation — Method

polaron_correlation(τ, a, params::OhmicBath)

Calculate the polaron transformed correlation function of Ohmic bath. a is the effective system bath coupling strength. It is the Hamming distance of two energy levels with respect to the system bath coupling operator.

QuantumAnnealingTools.γ — Method

γ(ω, params::OhmicBath)

Calculate Ohmic spectrum density, defined as a full Fourier transform on the bath correlation function.

QuantumAnnealingTools.S — Method

S(w, params::OhmicBath; atol=1e-7)

Calculate the Lamb shift of Ohmic spectrum. atol is the absolute tolerance for Cauchy principal value integral.

QuantumAnnealingTools.interpolate_spectral_density — Method

interpolate_spectral_density(ω_grid::AbstractRange{T}, params::OhmicBath) where T<:Number

Calculate the Ohmic bath spectral density S on grid ω_grid, and construct interpolation objects for it. A separate function for γ is also returned without doing interpolation.

Hybrid Ohmic Bath

QuantumAnnealingTools.HybridOhmicBath — Type

struct HybridOhmicBath

A hybrid noise model with both low and high frequency noise. The high frequency noise is characterized by Ohmic bath and the low frequence noise is characterized by the MRT width W.

W

MRT width (2π GHz)

ϵl

low spectrum reorganization energy (2π GHz)

ϵ

total reorganization energy (2π GHz)

η

strength of high frequency Ohmic bath

ωc

cutoff frequency

β

inverse temperature

width_h

half width at half maximum for high frequency Ohmic bath

width_l

half width at half maximum for low frequency bath

QuantumAnnealingTools.HybridOhmic — Method

HybridOhmic(W, η, fc, T)

Construct HybridOhmicBath object with parameters in physical units. W: MRT width (mK); η: interaction strength (unitless); fc: Ohmic cutoff frequency (GHz); T: temperature (mK).

QuantumAnnealingTools.polaron_correlation — Function

polaron_correlation(τ, bath::HybridOhmicBath, a=1)

Calculate polaron transformed correlation function of HybridOhmicBath 'bath' at time 'τ' with relative strength a. The effective strength will be $a * W^2$ and $a * η$.

QuantumAnnealingTools.Gₕ — Function

Gₕ(ω, bath::HybridOhmicBath, a=1)

High frequency noise spectrum of the HybridOhmicBath bath with relative strength a.

QuantumAnnealingTools.Gₗ — Function

Gₗ(ω, bath::HybridOhmicBath, a=1)

Low frequency noise specturm of the HybridOhmicBath bath with relative strength a.