Bath Module

Ohmic bath

OpenQuantumBase.OhmicBath — Type

OhmicBath

Ohmic bath object to hold a particular parameter set.

Fields

η – strength.

ωc – cutoff frequence.

β – inverse temperature.

OpenQuantumBase.Ohmic — Function

Ohmic(η, fc, T)

Construct OhmicBath from parameters with physical unit: η–unitless interaction strength; fc–cutoff frequency in GHz; T–temperature in mK.

Hybrid Ohmic bath

OpenQuantumBase.HybridOhmicBath — Type

struct HybridOhmicBath <: OpenQuantumBase.AbstractBath

A hybrid noise model with both low and high frequency noise. The high frequency noise is characterized by Ohmic bath and the low frequence noise is characterized by the MRT width W.

W

MRT width (2π GHz)

ϵl

low spectrum reorganization energy (2π GHz)

η

strength of high frequency Ohmic bath

ωc

cutoff frequency

β

inverse temperature

OpenQuantumBase.HybridOhmic — Function

HybridOhmic(W, η, fc, T)

Construct HybridOhmicBath object with parameters in physical units. W: MRT width (mK); η: interaction strength (unitless); fc: Ohmic cutoff frequency (GHz); T: temperature (mK).

Spin-fluctuator

OpenQuantumBase.EnsembleFluctuator — Type

struct EnsembleFluctuator{T} <: OpenQuantumBase.StochasticBath

An ensemble of random telegraph noise.

f

A list of RTNs

Others

OpenQuantumBase.CustomBath — Type

mutable struct CustomBath <: OpenQuantumBase.AbstractBath

An custum bath object defined by the two-point correlation function and the corresponding spectrum.

cfun

correlation function

γ

spectrum

OpenQuantumBase.CorrelatedBath — Type

mutable struct CorrelatedBath <: OpenQuantumBase.AbstractBath

CorrelatedBath defines a correlated bath type by the matrices of its two-point correlation functions and corresponding spectrums.

cfun

correlation function

γ

spectrum

inds

bath correlator idx

Common interface

OpenQuantumBase.correlation — Method

correlation(τ, bath)

Calculate the two point correlation function $C(τ)$ of bath.

OpenQuantumBase.γ — Method

γ(ω, bath)

Calculate spectral density $γ(ω)$ of bath.

OpenQuantumBase.spectrum — Method

spectrum(ω, bath)

Calculate spectral density $γ(ω)$ of bath.

OpenQuantumBase.S — Method

S(w, bath; kwargs...)

Calculate the Lamb shift of bath. All the keyword arguments of quadgk function is supported.