Bath Module

Ohmic bath

QTBase.OhmicBath — Type

OhmicBath

Ohmic bath object to hold a particular parameter set.

Fields

η – strength.

ωc – cutoff frequence.

β – inverse temperature.

QTBase.Ohmic — Function

Ohmic(η, fc, T)

Construct OhmicBath from parameters with physical unit: η–unitless interaction strength; fc–cutoff frequency in GHz; T–temperature in mK.

Common interface

QTBase.correlation — Method

correlation(τ, bath)

Calculate the two point correlation function $C(τ)$ of bath.

QTBase.γ — Method

γ(ω, bath)

Calculate spectrum density $γ(ω)$ of bath.

QTBase.spectrum — Method

spectrum(ω, bath)

Calculate spectrum density $γ(ω)$ of bath.

QTBase.S — Method

S(w, bath; atol)

Calculate the Lamb shift of Ohmic spectrum. atol is the absolute tolerance for Cauchy principal value integral.