Hamiltonian

Hamiltonian types

QTBase.DenseHamiltonian — Type

struct DenseHamiltonian{T<:Number} <: AbstractDenseHamiltonian{T<:Number}

Defines a time dependent Hamiltonian object with dense Matrices.

Fields

f

List of time dependent functions

m

List of constant matrices

u_cache

Internal cache

size

Size

EIGS

Eigen decomposition routine

QTBase.DenseHamiltonian — Method

DenseHamiltonian(funcs, mats; unit, EIGS)

Constructor of DenseHamiltonian object. funcs and mats are a list of time dependent functions and the corresponding matrices. unit is the unit one – :h or :ħ. The mats will be scaled by $2π$ if unit is :h.

EIGS is the initializer for the eigen decomposition routine for the Hamiltonian. It should return a function of signature: (H, s, lvl) -> (w, v). The default method EIGEN_DEFAULT will use LAPACK routine.

QTBase.SparseHamiltonian — Type

struct SparseHamiltonian{T<:Number} <: AbstractSparseHamiltonian{T<:Number}

Defines a time dependent Hamiltonian object with sparse Matrices.

Fields

f

List of time dependent functions

m

List of constant matrices

u_cache

Internal cache

size

Size

EIGS

Eigen decomposition routine

QTBase.SparseHamiltonian — Method

SparseHamiltonian(funcs, mats; unit, EIGS)

Constructor of SparseHamiltonian object. funcs and mats are a list of time dependent functions and the corresponding matrices. unit is the unit one – :h or :ħ. The mats will be scaled by $2π$ is unit is :h.

EIGS is the initializer for the eigen decomposition routine for the Hamiltonian. It should return a function of signature: (H, s, lvl) -> (w, v). The default method EIGEN_DEFAULT will use LAPACK routine.