Solver

Open System Solvers

QuantumAnnealingTools.jl provides several open quantum system solvers:

Redfield equation

The interface to Redfield equation is:

Basic usage

The basic usage of AME solve is

QuantumAnnealingTools.solve_redfield — Method.

function solve_redfield(A::Annealing, tf::Real, unitary; span_unit::Bool = false, kwargs...)

Solve the time dependent Redfield equation for Annealing defined by A with total annealing time tf.

...

Arguments

A::Annealing: the Annealing object.

tf::Real: the total annealing time.

unitary : precalculated unitary of close system evolution.

span_unit::Bool=false: flag variable which, when set to true, informs the solver to work with time in physical unit.

kwargs : other keyword arguments supported by DifferentialEquations.jl

...

source

Adiabatic master equation(AME)

A good reference to adiabatic master equation is Albash, Tameem, et al. QuantumAnnealingTools.jl provides the following interfaces for AME solver:

Basic usage

The basic usage of AME solver is

QuantumAnnealingTools.solve_ame — Method.

function solve_ame(
    A::Annealing,
    tf::Real;
    span_unit::Bool = false, ω_hint = [], lvl = size(A.H, 1), kwargs...,
)

Solve the adiabatic master equation for Annealing defined by A with total annealing time tf.

...

Arguments

A::Annealing: the Annealing object.

tf::Real: the total annealing time.

span_unit::Bool=false: flag variable which, when set to true, informs the solver to work with time in physical unit.

ω_hint=[] : grid for precalculating the lambshift; skip the precalculation if empty.

lvl::Int = size(A.H, 1) : number of levels to keep.

kwargs : other keyword arguments supported by DifferentialEquations.jl

...

source

Parallelization

solve_ame also allows one to perform parallel AME simulations for the same Annealing process for different total annealing time:

QuantumAnnealingTools.solve_ame — Method.

function solve_ame(
    A::Annealing,
    tf::Vector{T},
    alg,
    para_alg = EnsembleSerial();
    output_func = (sol, i) -> (sol, false),
    span_unit::Bool = false,
    ω_hint = [],
    lvl::Int = size(A.H, 1),
    kwargs...,
) where {T<:Real}

Solve the adiabatic master equation for Annealing defined by A with different total annealing time tf in parallel.

...

Arguments

A::Annealing: the Annealing object.

tf::Vector{T<:Real}: the total annealing times.

alg: solver algorithm

para_alg=EnsembleSerial(): ensemble algorithm.

output_func=(sol, i)->(sol, false): The function determines what is saved from the solution to the output array. Defaults to saving the solution itself. The output is (out,rerun) where out is the output and rerun is a boolean which designates whether to rerun.

span_unit::Bool=false: flag variable which, when set to true, informs the solver to work with time in physical unit.

ω_hint=[] : grid for precalculating the lambshift; skip the precalculation if empty.

lvl::Int = size(A.H, 1) : number of levels to keep.

kwargs : other keyword arguments supported by DifferentialEquations.jl

...

source

The key word argument para_alg specifies what EnsembleAlgorithm to use for Parallelization.

Additional solver options

The extra keyword arguments kwargs will be directly passed to ODE solve interface. So every solver option is supported. However, because of the current implementation, several keyword arguments are strongly recommended:

alg=Tsit5(): This set the algorithm to Tsitouras 5/4 Runge-Kutta method, which performs well for most problems. You can find a list of the available solver algorithms here. The implicit methods are not supported.

saveat: This argument denotes specific times to save the solution at, during the solving phase. It overrides the save_everystep keyword argument.

tstops: This argument denotes extra times that the timestepping algorithm must step to. It will increase the accuracy if the algorithm is instructed to step to the know singularities of the problem.

save_everystep: Whether to save the result at every step. You can save the memory by setting it to false.

Examples

An tutorial notebook for solving both Redfield and adiabatic master equation can be found here.